+++ craytraverse/renderer/__init__.py
@@ -1,0 +1,16 @@
+# -*- coding: utf-8 -*-
+# Copyright (c) 2020 Stephen Wasilewski, HSLU and EPFL
+# =======================================================================
+# This Source Code Form is subject to the terms of the Mozilla Public
+# License, v. 2.0. If a copy of the MPL was not distributed with this
+# file, You can obtain one at http://mozilla.org/MPL/2.0/.
+# =======================================================================
+
+"""renderer objects"""
+
+__all__ = ['Rtrace', 'Rcontrib']
+
+from craytraverse.renderer.renderer import Renderer
+from craytraverse.renderer.rtrace import Rtrace
+from craytraverse.renderer.rcontrib import Rcontrib
+